+++ pkg/analyzer/visitors.go
@@ -125,6 +125,25 @@ func (visitors *Visitors) TotalPageViews(filter *Filter) (int, error) {
 	return total, nil
 }
 
+// TotalSessions returns the total number of sessions.
+func (visitors *Visitors) TotalSessions(filter *Filter) (int, error) {
+	filter = visitors.analyzer.getFilter(filter)
+	f := &Filter{
+		ClientID: filter.ClientID,
+		Timezone: filter.Timezone,
+		From:     filter.From,
+		To:       filter.To,
+	}
+	q, args := f.buildQuery([]Field{FieldSessions}, nil, nil)
+	total, err := visitors.store.GetTotalSessionStats(q, args...)
+
+	if err != nil {
+		return 0, err
+	}
+
+	return total, nil
+}
+
 // TotalVisitorsPageViews returns the total visitor count and number of page views including the growth.
 func (visitors *Visitors) TotalVisitorsPageViews(filter *Filter) (*model.TotalVisitorsPageViewsStats, error) {
 	filter = visitors.analyzer.getFilter(filter)