+++ pkg/db/client_mock.go
@@ -192,6 +192,11 @@ func (client *ClientMock) GetTotalPageViewStats(string, ...any) (int, error) {
 	return 0, nil
 }
 
+// GetTotalSessionsStats implements the Store interface.
+func (client *ClientMock) GetTotalSessionStats(string, ...any) (int, error) {
+	return 0, nil
+}
+
 // SelectTimeSpentStats implements the Store interface.
 func (client *ClientMock) SelectTimeSpentStats(pkg.Period, string, ...any) ([]model.TimeSpentStats, error) {
 	return nil, nil